pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "clify"
description = "OpenAPI仕様からCLIを自動生成するツール"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
requires-python = ">=3.10.0"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
    "requests>=2.25.0",
]
dynamic = ["version"]

[project.scripts]
clify = "clify.cli:main"

[tool.setuptools.package-dir]
clify = "src/clify"

[tool.setuptools_scm]
write_to = "src/clify/_version.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
